% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllTheFunctions.R
\name{calculateVIFandRemoveVariables}
\alias{calculateVIFandRemoveVariables}
\title{Calculate variance inflation factor (VIF) and remove variables with VIF > threshold value}
\usage{
calculateVIFandRemoveVariables(response, raster_stack, thresh = 10,
  trace = T, ...)
}
\arguments{
\item{response}{A data frame containing the response data including number of positive cases (n_positive), number of individuals examined (examined), and point locations (longitude and latitude).}

\item{raster_stack}{A collection of RasterLayer objects with the same spatial extent and resolution (These are the environmental covariates).}

\item{thresh}{The threshold value used for retaining variables, Default: 10.}

\item{trace}{A logical argument indicating if text output is returned as the stepwise selection progresses, Default: T.}

\item{...}{Additional arguments passed to `lm'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The function checks for collinearity between variables and performs stepwise VIF selection. Source: https://beckmw.wordpress.com/2013/02/05/collinearity-and-stepwise-vif-selection/.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Marcus W. Beck
}
