% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllTheFunctions.R
\name{makeMeshSPDE}
\alias{makeMeshSPDE}
\title{Make a mesh, build an spde model, and a projector matrix}
\usage{
makeMeshSPDE(response, control = list(convex = NULL, concave = NULL, max_edge
  = NULL, cutoff = NULL), plot_mesh = FALSE)
}
\arguments{
\item{response}{A data frame containing the response data including number of positive cases (n_positive), number of individuals examined (examined), and point locations (longitude and latitude).}

\item{control}{Parameters controlling inla.nonconvex.hull and inla.mesh.2d, including convex, concave, max_edge, and cutoff.}

\item{plot_mesh}{A logical argument indicating if mesh is to be plotted, Default: FALSE.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Construct a nonconvex boundary for a set of points. Create a triangle mesh based on initial point locations, specified or automatic boundaries, and mesh quality parameters. Construct observation/prediction weight matrices for models. Create an inla.spde2 model object for a Matern model.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
