% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllTheFunctions.R
\name{predictionOnAGrid}
\alias{predictionOnAGrid}
\title{Prediction of the response on a grid/raster}
\usage{
predictionOnAGrid(response, finalmodel, A_mat, spde, mesh,
  family = "binomial", raster_stack, nsamp, int.strategy,
  write_posterior = TRUE)
}
\arguments{
\item{response}{A data frame containing the response data including number of positive cases (n_positive), number of individuals examined (examined), and point locations (longitude and latitude).}

\item{finalmodel}{An object returned from the function findModelWithSmallestDIC.}

\item{A_mat}{An observation/prediction weight matrix returned from makeMeshSPDE.}

\item{spde}{An inla.spde2 model object for a Matern model returned from makeMeshSPDE,}

\item{mesh}{A triangle mesh created based on initial point locations and returned from makeMeshSPDE.}

\item{family}{A string indicating the likelihood family, Default: `binomial'.}

\item{raster_stack}{A collection of RasterLayer objects with the same spatial extent and resolution.}

\item{nsamp}{Number of samples to draw from an approximated posterior of a fitted model. Make nsamp >= 100 in order to compute mean, sd, IQR, and 95\% CI of posterior samples.}

\item{int.strategy}{Character. The integration strategy to use; one of `auto', `ccd', `grid', `eb' (empirical bayes), `user' or `user.std'.}

\item{write_posterior}{A logical argument indicating if posterior realizations are to be written into raster files, Default: TRUE.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The function predicts the response on target locations where data are not observed using posterior distributions.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
